Rename token validation file to fix app crashing
