+++ express-back-end/src/api/auth/middlewares/index.js
@@ -1,7 +1,7 @@
 const ensureLoggedIn = require('./ensureLoggedIn');
-const tokenValidation = require('./tokenValidation');
+const checkTokenSetUser = require('./checkTokenSetUser');
 
 module.exports = {
   ensureLoggedIn,
-  tokenValidation,
+  checkTokenSetUser,
 };
